+++ test_files/guppy_optimization/nested/nested.py
@@ -1,7 +1,7 @@
 # /// script
 # requires-python = ">=3.13"
 # dependencies = [
-#     "guppylang >=0.21.3",
+#     "guppylang >=0.21.6",
 # ]
 # ///