+++ tket/src/serialize/pytket/extension/core.rs
@@ -10,7 +10,7 @@ use crate::serialize::pytket::decoder::{
     DecodeStatus, LoadedParameter, PytketDecoderContext, TrackedBit, TrackedQubit,
 };
 use crate::serialize::pytket::extension::PytketDecoder;
-use crate::serialize::pytket::opaque::{OpaqueSubgraphPayload, OPGROUP_OPAQUE_HUGR};
+use crate::serialize::pytket::opaque::OpaqueSubgraphPayload;
 use crate::serialize::pytket::{DecodeInsertionTarget, DecodeOptions, PytketDecodeError};
 use hugr::builder::Container;
 use hugr::extension::prelude::{bool_t, qb_t};
@@ -36,7 +36,7 @@ impl PytketDecoder for CoreDecoder {
         qubits: &[TrackedQubit],
         bits: &[TrackedBit],
         params: &[LoadedParameter],
-        opgroup: Option<&str>,
+        _opgroup: Option<&str>,
         decoder: &mut PytketDecoderContext<'h>,
     ) -> Result<DecodeStatus, PytketDecodeError> {
         match &op {
@@ -44,11 +44,10 @@ impl PytketDecoder for CoreDecoder {
                 op_type: PytketOptype::Barrier,
                 data: Some(payload),
                 ..
-            } if opgroup == Some(OPGROUP_OPAQUE_HUGR) => {
-                let payload =
-                    OpaqueSubgraphPayload::load_str(payload, decoder.extension_registry())?;
-                decoder.insert_subgraph_from_payload(qubits, bits, params, &payload)
-            }
+            } => match OpaqueSubgraphPayload::load_str(payload, decoder.extension_registry()) {
+                Ok(payload) => decoder.insert_subgraph_from_payload(qubits, bits, params, &payload),
+                _ => Ok(DecodeStatus::Unsupported),
+            },
             PytketOperation {
                 op_type: PytketOptype::CircBox,
                 op_box: